requirements.txt
# --------- torch --------- #
torch==1.9.0
torchvision
pytorch-lightning==1.4.8

# --------- hydra --------- #
hydra-core==1.1.0
hydra-colorlog==1.1.0
hydra-optuna-sweeper==1.1.0

# --------- loggers --------- #
wandb

# --------- others --------- #
rich            # beautiful text formatting in terminal
scipy
pyro-ppl
